pyproject.toml
[build-system]
requires = ["setuptools>=58", "wheel>=0.37"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-clean"
version = "0.1.0"
description = "ButlerSpec-governed cleanup assistant CLI skeleton"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    "pydantic>=2.6,<3",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6"]

[project.scripts]
"ai-clean" = "ai_clean.cli:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I"]
# E = pycodestyle errors, F = pyflakes, I = isort (ruff can handle imports too!)
